+++ entities/adhesion_paiement/classes.php
@@ -96,7 +96,7 @@ class AmapressAdhesion_paiement extends Amapress_EventBase {
 					'key'     => 'amapress_contrat_paiement_date',
 					'value'   => Amapress::add_days( $date, - 15 ),
 					'compare' => '>=',
-					'type'    => 'INT'
+					'type'    => 'NUMERIC'
 				),
 				array(
 					'key'     => 'amapress_contrat_paiement_adhesion',
@@ -122,7 +122,7 @@ class AmapressAdhesion_paiement extends Amapress_EventBase {
 					'key'     => 'amapress_contrat_paiement_date',
 					'value'   => array( $start_date, $end_date ),
 					'compare' => 'BETWEEN',
-					'type'    => 'INT'
+					'type'    => 'NUMERIC'
 				),
 			),
 			$order );
